data_extractor_api/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
Requests==2.32.3
